add ovc_machines_aio first refactoring
ovc_machines_aio is one new part of the more flexible ovc setup script.

The previous workflow was:
 - for each machines:
  - creating machine on ms1
  - installing and configuring that node

This script will be used together with 'ovc_configure_aio' and will use
this new workflow:
 - for each machines:
  - creating machine to the remote target (ms1, docker, ...)

 - for each machines:
  - configuring that machine

This new workflow enable us to use a common script for configuration and
a flexible script for machine creation and management. This separation
allow us to use multiple machines type with the same setup deployment.

This commit only push the first version of the new ovc_machines_aio
stuff which is in unstable state. The previous workflow is untouched and
is still used for production.